+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## Next
 
+### Calendars
+
+- `TextInput` now automatically changes `iconLeft` into a `FlatButton` if `onClickLeft` is set. Same for right side.
+
+#### minDate and maxDate props
+
+New props `minDate` and `maxDate` in `Calendar` and all calendar input components.
+
+Both are optional, and both are strings in format `2021-11-23`.
+
+When set, all dates before (or after) will be disabled in the calendar.
+
 ### Heading and Text
 
 - Added `wordBreak` prop.